requirements.txt
google-generativeai==0.7.2
python-dotenv==1.0.1
streamlit==1.38.0
requests==2.32.3
